♻️ refactor: move hooks
