+++ src/features/ChatInput/STT/browser.tsx
@@ -1,11 +1,39 @@
+import { SpeechRecognitionOptions, useSpeechRecognition } from '@lobehub/tts/react';
+import { isEqual } from 'lodash';
 import { memo, useCallback, useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import { SWRConfiguration } from 'swr';
 
-import { useBrowserSTT } from '@/hooks/useSTT';
 import { useChatStore } from '@/store/chat';
+import { useGlobalStore } from '@/store/global';
+import { settingsSelectors } from '@/store/global/selectors';
+import { useSessionStore } from '@/store/session';
+import { agentSelectors } from '@/store/session/selectors';
 import { ChatMessageError } from '@/types/message';
 import { getMessageError } from '@/utils/fetch';
+
 import CommonSTT from './common';
+
+interface STTConfig extends SWRConfiguration {
+  onTextChange: (value: string) => void;
+}
+
+const useBrowserSTT = (config: STTConfig) => {
+  const ttsSettings = useGlobalStore(settingsSelectors.currentTTS, isEqual);
+  const ttsAgentSettings = useSessionStore(agentSelectors.currentAgentTTS, isEqual);
+  const locale = useGlobalStore(settingsSelectors.currentLanguage);
+
+  const autoStop = ttsSettings.sttAutoStop;
+  const sttLocale =
+    ttsAgentSettings?.sttLocale && ttsAgentSettings.sttLocale !== 'auto'
+      ? ttsAgentSettings.sttLocale
+      : locale;
+
+  return useSpeechRecognition(sttLocale, {
+    ...config,
+    autoStop,
+  } as SpeechRecognitionOptions);
+};
 
 const BrowserSTT = memo<{ mobile?: boolean }>(({ mobile }) => {
   const [error, setError] = useState<ChatMessageError>();
@@ -71,19 +99,19 @@ const BrowserSTT = memo<{ mobile?: boolean }>(({ mobile }) => {
   }, [start]);
 
   return (
-    <CommonSTT 
+    <CommonSTT
       desc={desc}
       error={error}
       formattedTime={formattedTime}
-      handleCloseError={handleCloseError} 
+      handleCloseError={handleCloseError}
       handleRetry={handleRetry}
-      handleTriggerStartStop={handleTriggerStartStop} 
+      handleTriggerStartStop={handleTriggerStartStop}
       isLoading={isLoading}
       isRecording={isRecording}
-      mobile={mobile} 
+      mobile={mobile}
       time={time}
-      />
-  )
+    />
+  );
 });
 
 export default BrowserSTT;
